fixed "scrolling speed" config option working in reverse.
This change is basically a UI change: the "slow" and "fast" labels have been reverse so that it is clear that lower values are "slower" (make the delay loop longer) and higher numbers are "faster" (decrese the delay loop). Also the number of steps were changed - previously any speed value setting of 6 or higher would translate to "as fast as possible", which means there were effectively only 6 setting and none slower than 50ms loop delay.

So I have also changed the range from "1 through 10" to "-4 through 6" - so now:
- It is clear that the highest speed is 6
- The default value ("2") means the same, as well as any other valid (6 or less) value
- If the installation is configured to a value higher than 6, it will continue to behave as 6, but will be clamped down to 6 when the configurtion panel is opened.
- values lower than 1 are now possible and mean slowing the scroll down to a loop delay of 100ms. I am not sure that this is the slowest speed people will want, but it is 2.5 times slower than the default, so... maybe?